+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { ThemeProvider } from "@mui/material";
 import App from "./App";
-import customTheme from "./customizing/customTheme";
+import customTheme from "./theme/customTheme";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
